+++ tests/loops_kernels/loop_general_dec/loop_general_dec_hands.cl
@@ -13,9 +13,9 @@ void loop_general(__global int *data, int from, int to, int x)
         do {
             var3 = x * var4;
             var4 = (ulong)var5 + (var3);
-            var5 = var5 + (ulong)(-1);
-        } while (!((var5 + (ulong)(-1)) <= to));
-        var2 = (var3) + (ulong)(var5 + (ulong)(-1)) + 1;
+            var5 = var5 + (-1);
+        } while (!((var5 + (-1)) <= to));
+        var2 = (var3) + (ulong)(var5 + (-1)) + 1;
     }
     data[get_global_id(0)] = var2;
 }
