+++ tests/loops_kernels/simple_loop_kernels/simple_loop_kernels_dcmpl.cl
@@ -44,11 +44,11 @@ void loop_kernel_1(__global uint *data, uint x, uint y, uint unrollingBreaker)
             var0 = *var12;
             var8 = var8 + 4 / 4;
             var9 = var9 + 0 / 4;
-            var10 = var10 + (ulong)(-1);
+            var10 = var10 + (-1);
             var11 = var0 + var11;
             *var12 = var11;
             var11 = (var11) * x;
-        } while (!((int)(var10 + (ulong)(-1)) == 0));
+        } while (!((int)(var10 + (-1)) == 0));
     }
 }
 
@@ -71,10 +71,10 @@ void loop_kernel_2(__global uint *data, uint x, uint y, uint unrollingBreaker)
             var0 = *var12;
             var8 = var8 + (-4) / 4;
             var9 = var9 + (-1) / 4;
-            var10 = var10 + (ulong)(-1);
+            var10 = var10 + (-1);
             var11 = var0 + var11;
             *var12 = var11;
             var11 = (var11) * x;
-        } while (!((int)(var10 + (ulong)(-1)) == 0));
+        } while (!((int)(var10 + (-1)) == 0));
     }
 }
